put enum and jsonapi utils inside the Patreon::Utils module
